# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.CreateRegistroDiarioDTO:
    properties:
      conteudo:
        minLength: 5
        type: string
      data:
        type: string
      tipo:
        allOf:
        - $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.RegistroTipo'
        enum:
        - observacao
        - evento
        - aprendizado
        - tratamento
        - problema
        - colheita
        - crescimento
      titulo:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - conteudo
    - data
    - tipo
    - titulo
    type: object
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.PaginatedResponse:
    properties:
      data:
        items:
          type: integer
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.RegistroDiarioResponseDTO:
    properties:
      conteudo:
        type: string
      data:
        type: string
      id:
        type: integer
      tipo:
        $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.RegistroTipo'
      titulo:
        type: string
    type: object
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_entity.RegistroTipo:
    enum:
    - observacao
    - evento
    - aprendizado
    - tratamento
    - problema
    - colheita
    - crescimento
    type: string
    x-enum-varnames:
    - RegistroTipoObservacao
    - RegistroTipoEvento
    - RegistroTipoAprendizado
    - RegistroTipoTratamento
    - RegistroTipoProblema
    - RegistroTipoColheita
    - RegistroTipoCrescimento
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for the Cultivo app.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cultivo API
  version: "1.0"
paths:
...
  /api/v1/diarios-cultivo/{diario_id}/registros:
    get:
      consumes:
      - application/json
      description: Retorna uma lista paginada de todos os registros diários de um
        diário de cultivo específico
      parameters:
      - description: ID do Diário de Cultivo
        in: path
        name: diario_id
        required: true
        type: integer
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Limite de itens por página (padrão: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.RegistroDiarioResponseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Lista todos os registros diários de um diário de cultivo com paginação
      tags:
      - registros-diario
    post:
      consumes:
      - application/json
      description: Adiciona um novo registro diário a um diário de cultivo específico
      parameters:
      - description: ID do Diário de Cultivo
        in: path
        name: diario_id
        required: true
        type: integer
      - description: Dados para criação do registro diário
        in: body
        name: registroDiario
        required: true
        schema:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.CreateRegistroDiarioDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.RegistroDiarioResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Cria um novo registro diário para um diário de cultivo
      tags:
      - registros-diario
...
swagger: "2.0"